requirements.txt
# ===== Core =====
numpy==1.26.4
pandas==2.2.2
joblib==1.3.2
tqdm==4.66.2

# ===== Computer Vision =====
opencv-python==4.9.0.80
Pillow==10.2.0
ultralytics==8.1.0

# ===== Deep Learning (Pipeline B) =====
torch==2.1.2+cpu
torchvision==0.16.2+cpu
torchaudio==2.1.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# ===== Machine Learning =====
scikit-learn==1.3.2
xgboost==2.0.3

# ===== Visualization =====
matplotlib==3.8.2
seaborn==0.13.2

# ===== Web =====
Flask==3.0.2
Werkzeug==3.0.1
